+++ lib/diamond/src/mesh/collision/mod.rs
@@ -11,6 +11,9 @@ use ndarray::Axis;
 use rayon::iter::IndexedParallelIterator;
 use rayon::iter::IntoParallelIterator;
 
+#[cfg(test)]
+mod tests;
+
 #[derive(Debug)]
 pub struct CollisionMesh {
 	pub obstacles: MultiPolygon<f32>,
@@ -114,103 +117,4 @@ impl CollisionMesh {
 			Point::new(extremes.xmax.x(), extremes.ymax.y()),
 		)
 	}
-}
-
-#[test]
-fn compare_meshes_equal_trivially_different() {
-	let a = CollisionMesh {
-		obstacles: vec![polygon![
-			(x: 16.0, y: 32.0),
-			(x: 0.0, y: 0.0),
-			(x: 128.0, y: 0.0),
-			(x: 128.0, y: 16.0),
-			(x: 16.0, y: 16.0),
-			(x: 16.0, y: 32.0),
-		]]
-		.into(),
-	};
-
-	let b = CollisionMesh {
-		obstacles: vec![polygon![
-			(x: 20.0, y: 20.0),
-			(x: 10.0, y: 20.0),
-			(x: 10.0, y: 10.0),
-			(x: 20.0, y: 20.0),
-		]]
-		.into(),
-	};
-
-	assert_ne!(a, b);
-}
-
-#[test]
-fn compare_meshes_different_but_overlapping() {
-	let a = CollisionMesh {
-		obstacles: vec![polygon![
-			(x: 0.0, y: 0.0),
-			(x: 10.0, y: 0.0),
-			(x: 10.0, y: 10.0),
-			(x: 0.0, y: 10.0),
-			(x: 0.0, y: 0.0),
-		]]
-		.into(),
-	};
-
-	let b = CollisionMesh {
-		obstacles: vec![polygon![
-			(x: 0.0, y: 0.0),
-			(x: 20.0, y: 0.0),
-			(x: 20.0, y: 10.0),
-			(x: 0.0, y: 10.0),
-			(x: 0.0, y: 0.0),
-		]]
-		.into(),
-	};
-
-	assert_ne!(a, b);
-}
-
-#[test]
-fn compare_meshes_cycled_vertices_and_polygons() {
-	let a = CollisionMesh {
-		obstacles: vec![
-			polygon![
-				(x: 16.0, y: 32.0),
-				(x: 0.0, y: 0.0),
-				(x: 128.0, y: 0.0),
-				(x: 128.0, y: 16.0),
-				(x: 16.0, y: 16.0),
-				(x: 16.0, y: 32.0),
-			],
-			polygon![
-				(x: 10.0, y: 10.0),
-				(x: 20.0, y: 20.0),
-				(x: 10.0, y: 20.0),
-				(x: 10.0, y: 10.0),
-			],
-		]
-		.into(),
-	};
-
-	let b = CollisionMesh {
-		obstacles: vec![
-			polygon![
-				(x: 20.0, y: 20.0),
-				(x: 10.0, y: 20.0),
-				(x: 10.0, y: 10.0),
-				(x: 20.0, y: 20.0),
-			],
-			polygon![
-				(x: 0.0, y: 0.0),
-				(x: 128.0, y: 0.0),
-				(x: 128.0, y: 16.0),
-				(x: 16.0, y: 16.0),
-				(x: 16.0, y: 32.0),
-				(x: 0.0, y: 0.0),
-			],
-		]
-		.into(),
-	};
-
-	assert_eq!(a, b);
 }
